optimal_folder_structure
