+++ package.json
@@ -15,7 +15,8 @@
     "@mui/icons-material": "^5.14.19",
     "@mui/material": "^5.14.20",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.20.1"
   },
   "devDependencies": {
     "@types/react": "^18.2.37",
