+++ src/components/SideBar/index.jsx
@@ -1,0 +1,11 @@
+import {} from 'react'
+import Box from '@mui/material/Box'
+const SideBar = () => {
+  return (
+    <div className="side-bar">
+      <Box sx={{ width: '68px' }}>SideBar</Box>
+    </div>
+  )
+}
+
+export default SideBar
