+++ src/main.jsx
@@ -1,16 +1,19 @@
 import React from 'react'
-import ReactDOM from 'react-dom/client'
+import './index.css'
 import App from './App.jsx'
+import theme from './theme.js'
+import ReactDOM from 'react-dom/client'
 import CssBaseline from '@mui/material/CssBaseline'
 import { Experimental_CssVarsProvider as CssVarsProvider } from '@mui/material/styles'
-import './index.css'
-import theme from './theme.js'
+import { BrowserRouter } from 'react-router-dom'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <CssVarsProvider theme={theme}>
-      <CssBaseline />
-      <App />
+      <BrowserRouter>
+        <CssBaseline />
+        <App />
+      </BrowserRouter>
     </CssVarsProvider>
   </React.StrictMode>
 )
